+++ NEWS.md
@@ -1,5 +1,10 @@
 # Release Notes Version 0.3
 
+* [0.3.5] 
+    * released 2025-02-14 
+    * rebuild rsp_match_profile using data.table 
+    * added rsp_plot_match (rsp_match_profile plot) output... 
+
 * [0.3.4] 
     * released 2024-08-06 
     * expended rsp_dcast to handle species, species.id, profile,